block: bdev_filter
New feature allows extending the block-level capabilities. When
processing bio in the submit_bio_noacct() function, it is possible to
invoke a filter. This allows to pass a bio, complete it, or redirect
it to another block device.

Each filter has a name, a pointer to the filter context, and a structure
with pointers to callback functions. The name must be unique, because
only one filter with the specified name can be attached to a single
block device. For this purpose, the name of the module that implements
the filter is perfect. The filter context is the personal data of the
filter. It can be a pointer to some structure. If the personal data is
not needed, it can be NULL. Only two callback functions are implemented.
The submit_bio_cb() is called from the submit_bio_noacct() function.
The detach_cb() is called when the block device is disconnected and
allows to safely release the resources allocated for the filter
operation.

There can be multiple filters for each block device. The filters are
added to a linked list per device. The order in which these filters are
called is currently determined only by the order in which these filters
are created. To synchronize access to the linked list,
percpu_rw_semaphore is used. This allows to minimize costs for readers.

In order for the filter to work for each block device, and not for the
disk, the blk_partition_remap() function was moved from
submit_bio_checks() to submit_bio_noacct() and located after applying
the filters.

Signed-off-by: Sergei Shtepa <sergei.shtepa@veeam.com>