block: Block Device Filtering Mechanism
See the patch below for what I think would be useful.

 - this changes the operations to be the only globally registered
   structure - for that it grows a name, owner and entry field.
   Refcounting is done using the module refcount to prevent unloading.
 - the blkfilter structure is repurposed so that it is embedded into
   the driver (i.e. tracker) structure.  This allows the driver
   to easily find things and removes the need for the separate
   lookup list in blksnap

There is a behavior change in the detach path - as the code says
a freeze is not required this change drops it as that nicely
simplifies the code.  On the attach side the attach is failed
if we can't freeze the file system instead of just printing an
error.