asynchronous snapimage_submit_bio() implementation
The I/O unit handler may fall asleep on the semaphore when lock a chunk,
but copying is performed at the moment when all the data for processing
the I/O unit has already been prepared.
If all the necessary chunks are already in the cache, then the reading
is performed synchronously, without accessing the block devices.
If some data needs to be loaded from the original block device or
difference storage, then copying data and completing the I/O unit
is performed in the callback function.