block: Block Device Filtering Mechanism - using bio chaining
Also the image ctx would need a mempool allocation.  But I have an idea
how to eventually kill it using bio chaining.  But for that I first
looked into the chunk I/O code, so please take a look at the patch
below:  this removes the diff_io structure and moves the remaining
bits into struct chunk.  There is no more need for the bio_count
due to the use of bio chaining.  I might also be able to get rid of
the error and is_write fields as follow on, but it didn't really
seem worth it.