block: Block Device Filtering Mechanism - remove bio_list hacks
I think the overall structure looks really nice.  But and async submit
would be even nicer, and the bio_list hacks are probably not
acceptable upsteam anyway.

What do you think of the patch below that stashes away an iter to
copy into and the processes completes the original bio from the
final chunk load using the bi_remaining mechanism?