remove clear flag 'BIO_FILTERED' for bio
Clearing the 'BIO_FILTERED' flag so that the filters of other block
devices can also apply their own filter is of course good idea, but
unfortunately it does not work yet. In the case of DM, difficulties
arise.

At the moment, there is no need for such functionality. Maybe we can
bring back the flag cleanup when we need it. But do not forget to check
for compatibility with DM devices.