set current->blk_filter in diff_area_image_io_work()
Since reading and writing to the difference storage file is placed in
a separate thread (system worker), it is necessary to specify a filter
for it.

To prevent access to the file system when allocating memory during
snapshot images I/O, function calls memalloc_nofs_save() and
memalloc_nofs_restore() were added.

Fixed a bug in the negative branch of the algorithm in function
snapimage_submit_bio(). The value of current->blk_filter was not
restored when an I/O unit processing error occurred.