add config BLKSNAP_DIFF_BLKDEV
The BLKSNAP_DIFF_BLKDEV configuration allows to use block devices for
storing changes to difference storage. To work with regular files, a
virtual file system API is used. Lower-level functions are used to work
with block devices. The block device opens in exclusive access mode.

If this option is disabled, the difference storage uses the block
device as a regular file.