add chunk_load_failed()
The function chunk_io_ctx_free() can be called when processing an I/O
unit with an error code is completed. The reasons for the error are not
related to the operation of the module and the state of the snapshot.
For example, a I/O request by an offset that is not aligned along the
border of the block. So, this is not a reason to declare a snapshot
corrupted.