prevents adding inline encrypted devices to the snapshot
There is an opinion that the use of the blksnap module may violate the
security of encrypted data. The difference storage file may be located
on an unreliable disk or even network storage. To implement secure
compatibility with inline encrypted devices will require discussion of
algorithms and restrictions. For example, a restriction on the location
of the difference storage only in virtual memory might help. Currently,
there is no need for compatibility of the blksnap module and inline
encryption.