remove file_start_write()/file_end_write()
First of all, we don't need to block the superblock.
Secondly, locks are performed using functions percpu_down_read() and
percpu_up_read(). Locking and unlocking must be performed in the same
thread. And in our case, unblocking was performed in the workflow,
which led to an error "workqueue leaked lock or atomic".