block: difference storage implementation
The struct diff_storage manages the block device or file that are used
to store the data of the original block devices in the snapshot.
The difference storage is created one per snapshot and is used to store
data from all block devices.

The struct diff_area provides management of the difference blocks of
a single block device. Storing difference blocks, and reading them to
get a snapshot images.

The struct chunk describes the minimum data storage unit of the original
block device. Functions for working with these minimal blocks implement
algorithms for reading and writing blocks.

The struct diff_buffer describes the buffer in memory for a chunk.

Co-developed-by: Christoph Hellwig <hch@infradead.org>
Signed-off-by: Christoph Hellwig <hch@infradead.org>